Refactored out apiHooks, does not work.
